fixed a bug where :command action :Data was not correctly processed inside the :Data map.
the problem was inside the parsing strategy since all fields using ":" as separator were treated as packet fields regarding if they appeared after a Response:Follows or not.

now a special case is added and if the Response is Follows all fields no matter the syntax are treated as data. since ordering is not assured testing is required to validate the two standard Privilege and Action fields are still treated as meta.